+++ app/src/features/apiClient/screens/apiClient/apiClient.scss
@@ -3,7 +3,7 @@
   display: flex;
 
   .api-client-sidebar {
-    width: 320px;
+    width: 280px;
     flex: 0 0 auto;
     border-right: 1px solid var(--border-dark);
     background-color: var(--component-background);
