+++ app/src/features/apiClient/screens/apiClient/components/sidebar/apiClientSidebar.scss
@@ -25,11 +25,6 @@
     align-items: center;
   }
 
-  .api-client-sidebar-content {
-    display: flex;
-    height: 100%;
-  }
-
   .api-history-list {
     width: 100%;
     height: 100%;
