+++ app/src/features/apiClient/screens/apiClient/components/sidebar/apiClientSidebarTabs/apiClientSidebarTabs.scss
@@ -1,9 +1,31 @@
 .api-client-sidebar {
-  .api-client-secondary-sidebar {
+  .api-client-sidebar-tabs {
     flex-shrink: 0;
     height: 100%;
     background-color: var(--body-background);
-    border-right: 1px solid var(--border-dark);
+
+    .ant-tabs-nav {
+      margin-bottom: 0;
+      padding: 0 8px;
+
+      .ant-tabs-tab + .ant-tabs-tab {
+        margin-left: 16px;
+      }
+
+      .ant-tabs-tab {
+        .api-client-tab-label {
+          display: flex;
+          align-items: center;
+          gap: 4px;
+          font-size: var(--requestly-font-size-sm);
+
+          svg {
+            width: 12px;
+            height: 12px;
+          }
+        }
+      }
+    }
 
     .sidebar-item {
       line-height: 15.92px;
@@ -21,11 +43,6 @@
       justify-content: center;
       row-gap: 8px;
       color: var(--text-dark-gray); // TODO: update the variable
-
-      .item-icon svg {
-        width: 20px;
-        height: 20px;
-      }
 
       &:hover,
       &.active {
